+++ src/StoreKeeper/WooCommerce/B2C/Frontend/FrontendCore.php
@@ -4,8 +4,8 @@ namespace StoreKeeper\WooCommerce\B2C\Frontend;
 
 use StoreKeeper\WooCommerce\B2C\Frontend\Filters\OrderTrackingMessage;
 use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\AddressFormHandler;
-use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\CartHandler;
 use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\CategorySummaryHandler;
+use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\CustomerEmailHandler;
 use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\MarkdownHandler;
 use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\OrderHookHandler;
 use StoreKeeper\WooCommerce\B2C\Frontend\Handlers\ProductAddOnHandler;
@@ -54,7 +54,7 @@ class FrontendCore
         $seo = new StoreKeeperSeoHandler();
         $seo->registerHooks();
 
-        $cartHandler = new CartHandler();
+        $cartHandler = new CustomerEmailHandler();
         $cartHandler->registerHooks();
 
         $categorySummray = new CategorySummaryHandler();
